refactor: change dir structure according to surf standarts
