+++ lib/features/dash/presentation/screens/dash_screen_model.dart
@@ -1,0 +1,5 @@
+import 'package:elementary/elementary.dart';
+import 'package:flutter_template/features/dash/presentation/screens/dash_screen.dart';
+
+/// Model fol [DashScreen].
+class DashScreenModel extends ElementaryModel {}
