--- lib/features/dash/screen/dash_screen_model.dart
@@ -1,5 +1,0 @@
-import 'package:elementary/elementary.dart';
-import 'package:flutter_template/features/dash/screen/dash_screen.dart';
-
-/// Model fol [DashScreen].
-class DashScreenModel extends ElementaryModel {}
