+++ lib/features/info/presentation/screens/info_screen_model.dart
@@ -1,0 +1,5 @@
+import 'package:elementary/elementary.dart';
+import 'package:flutter_template/features/info/presentation/screens/info_screen.dart';
+
+/// Model fol [InfoScreen].
+class InfoScreenModel extends ElementaryModel {}
