--- lib/features/info/screen/info_screen_model.dart
@@ -1,5 +1,0 @@
-import 'package:elementary/elementary.dart';
-import 'package:flutter_template/features/info/screen/info_screen.dart';
-
-/// Model fol [InfoScreen].
-class InfoScreenModel extends ElementaryModel {}
