+++ lib/features/temp/data/repositories/template_repo.dart
@@ -1,0 +1,4 @@
+import 'package:flutter_template/features/temp/domain/repositories/i_template_repo.dart';
+
+/// Implementation of Template repo.
+class TemplateRepo implements ITemplateRepo {}
