+++ lib/features/temp/domain/repositories/i_template_repo.dart
@@ -1,0 +1,2 @@
+/// Interface of the template repo.
+abstract interface class ITemplateRepo {}
