+++ test/features/debug/screens/debug_screen/debug_screen_model_test.dart
@@ -4,8 +4,8 @@ import 'package:flutter_template/config/app_config.dart';
 import 'package:flutter_template/config/environment/environment.dart';
 import 'package:flutter_template/config/urls.dart';
 import 'package:flutter_template/features/common/service/theme/theme_service.dart';
-import 'package:flutter_template/features/debug/screens/debug_screen/debug_screen_model.dart';
-import 'package:flutter_template/features/debug/screens/debug_screen/debug_screen_widget_model.dart';
+import 'package:flutter_template/features/debug/presentation/screens/debug_screen/debug_screen_model.dart';
+import 'package:flutter_template/features/debug/presentation/screens/debug_screen/debug_screen_widget_model.dart';
 import 'package:flutter_template/persistence/storage/config_storage/config_storage.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
